+++ src/imio/history/tests/test_documentbylineviewlet.py
@@ -41,19 +41,19 @@ class TestDocumentByLineViewlet(IntegrationTestCase):
         # not highlighted because '' is an ignored comment
         history = adapter.getHistory()
         self.assertFalse(self.viewlet.highlight_history_link())
-        self.assertTrue(history[0]['comments'] in adapter.ignorableHistoryComments())
+        self.assertTrue(history[-1]['comments'] in adapter.ignorableHistoryComments())
 
         # now 'publish' the doc and add a comment, last event has a comment
         self.wft.doActionFor(self.portal.doc, 'publish', comment='my publish comment')
         history = adapter.getHistory()
         self.assertTrue(self.viewlet.highlight_history_link())
-        self.assertFalse(history[0]['comments'] in adapter.ignorableHistoryComments())
+        self.assertFalse(history[-1]['comments'] in adapter.ignorableHistoryComments())
 
         # now test the 'you can not access this comment' message
         self.wft.doActionFor(self.portal.doc, 'retract', comment=HISTORY_COMMENT_NOT_VIEWABLE)
         history = adapter.getHistory()
         self.assertFalse(self.viewlet.highlight_history_link())
-        self.assertTrue(history[0]['comments'] in adapter.ignorableHistoryComments())
+        self.assertTrue(history[-1]['comments'] in adapter.ignorableHistoryComments())
 
         # test that it works if no history
         # it is the case if we changed used workflow